Fixed a small navigation issue where scraper would not properly exit after rom found.

+++ static/build/.tmp_update/script/scraper/menu.sh
@@ -329,6 +329,7 @@ Launch_Scraping ()
             /mnt/SDCARD/.tmp_update/script/scraper/scrap_retroarch.sh $CurrentSystem "$onerom"
         fi
     fi
+    if [ -f "$romimage" ] && ! [ "$onerom" = "" ] ; then  echo exiting $romimage ;  exit;fi;
 
 	rm -f /tmp/stay_awake
     exit
